requirements.txt
django
djangorestframework
mysqlclient
pillow
django-filter==2.4.0
dj-rest-auth